fixed issue with marking habit done
Before I used to check if more than 1 day passed but that was incorrect.

Now I am just checking if the last time the habit was completed was before today since it's still supposed to show up even if less than a day has passed as long as its a different day that is subscribed to by the user.